+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/proxy/relay/RelayConnectHandler.java
@@ -116,6 +116,7 @@ public abstract class RelayConnectHandler<T> extends SimpleChannelInboundHandler
                                           mockServerCtx.channel().pipeline();
 
                                       if (isSslEnabledDownstream(proxyClientCtx.channel())) {
+                                        log.info("Adding SSL Handler in RelayConnectHandler.channelRead");
                                         pipelineToMockServer.addLast(
                                             server
                                                 .getClientSslContextFactory()